pyproject.toml
[tool.poetry]
name = "rteap-app"
version = "0.1.0"
description = ""
authors = ["ymuza <yamil.muza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.1.0"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "4.0.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
